server/backend/requirements.txt
paho-mqtt>=2.1.0
stmpy>=0.7.0